feat: mobile screens, pending result sheet, avatar upload, polish
Mobile (Expo/React Native):
- Full screen suite: Discover (infinite scroll, location/sport filters),
  My Games (upcoming/past split, live badge), Create Game (full form + GPS),
  Profile (sports, ELO, coach profiles, edit modal), Session Detail
  (join/leave/cancel/complete/report/confirm/dispute), Public Profile
- PendingResultSheet: transparent Modal + Animated spring + PanResponder
  drag-to-dismiss, sport-coloured accent strip, halo icon, session card
  with left border, player count row, press animation on CTA
- My Games amber banner with Trophy + ChevronRight in strict flex row
- pendingSheetStore (Zustand) for cross-component sheet control
- Avatar component: base64 image or AvatarInitials fallback
- Profile picture upload on register + edit profile modal (expo-image-picker)
- Dockerfile.dev: copy scripts/ before npm install so postinstall succeeds
- NativeWind patch script, LocationPicker, ConflictModal, places lib

Backend (Spring Boot):
- Game sessions: create, join/leave, status transitions (OPEN→FULL→IN_PROGRESS→COMPLETED),
  team assignment, result reporting + confirmation/dispute flow, ELO recalculation
- Pending results endpoint with participant count and location name
- Avatar data column (V17 migration), PATCH /me/profile endpoint
- Register accepts optional avatarData field
- Flyway migrations V13–V17 (Google place fields, IN_PROGRESS status,
  ELO decimal, result votes table, avatar column)
- ResultVote entity, SessionStatusScheduler, EloTransactionalService

+++ backend/backend/src/main/java/com/Levelmate/backend/elo/dto/EloHistoryPageResponse.java
@@ -3,7 +3,7 @@ package com.Levelmate.backend.elo.dto;
 import java.util.List;
 
 public record EloHistoryPageResponse(
-        Integer currentElo,
+        Double currentElo,
         int gamesPlayed,
         List<EloHistoryResponse> history,
         long totalElements,
